map객체 받아옴
