+++ app/src/main/AndroidManifest.xml
@@ -4,6 +4,8 @@
     package="com.maratangsoft.mykaraokebook">
 
     <uses-permission android:name="android.permission.INTERNET"/>
+    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION"/>
+    <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION"/>
 
     <application
         android:allowBackup="true"
