+++ app/src/main/java/com/maratangsoft/mykaraokebook/LocationFragment.kt
@@ -1,0 +1,37 @@
+package com.maratangsoft.mykaraokebook
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.maratangsoft.mykaraokebook.databinding.FragmentLocationBinding
+import com.naver.maps.map.MapFragment
+import com.naver.maps.map.NaverMap
+import com.naver.maps.map.OnMapReadyCallback
+
+class LocationFragment : Fragment(), OnMapReadyCallback {
+    private lateinit var binding: FragmentLocationBinding
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
+        binding = FragmentLocationBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        binding.btnSetting.setOnClickListener { startActivity(Intent(requireActivity(), SettingActivity::class.java)) }
+
+        val fm = requireActivity().supportFragmentManager
+        val map = fm.findFragmentById(R.id.map) as MapFragment?
+            ?: MapFragment.newInstance().also {
+                fm.beginTransaction().add(R.id.map, it).commit()
+            }
+        map.getMapAsync(this)
+    }
+
+    override fun onMapReady(p0: NaverMap) {
+
+    }
+}
